♻️ refactor(_c): consolidate CSS under css/
The CSS engines sat in three subsystem directories: the minifier under
serialize/, the selector matcher and css_to_xpath under query/css/, and
the CSSOM cascade under cssom/. Group them into one _c/css/ tree with
subdirs named for what each does: minify/, select/, cssom/.

The move changes no code generation. Includes resolve from the _c/ root,
so relocating a header only rewrites the path in each #include line, and
the preprocessor drops that path before the token stream. Every affected
translation unit produces byte-identical cc -E -P output and byte-identical
-O3 disassembly before and after, so LTO and PGO inlining and the CodSpeed
instruction counts hold.
