+++ meson.build
@@ -63,7 +63,7 @@ py.extension_module(
         'src/turbohtml/_c/serialize/lossless.c',
         'src/turbohtml/_c/serialize/c14n.c',
         'src/turbohtml/_c/serialize/minify.c',
-        'src/turbohtml/_c/serialize/css.c',
+        'src/turbohtml/_c/css/minify/css.c',
         'src/turbohtml/_c/serialize/markdown.c',
         'src/turbohtml/_c/serialize/text.c',
         'src/turbohtml/_c/extract/readability.c',
@@ -75,8 +75,8 @@ py.extension_module(
         'src/turbohtml/_c/serialize/js/mangle.c',
         'src/turbohtml/_c/serialize/js/minify.c',
         'src/turbohtml/_c/serialize/js/lexdump.c',
-        'src/turbohtml/_c/query/css/selector.c',
-        'src/turbohtml/_c/query/css/to_xpath.c',
+        'src/turbohtml/_c/css/select/selector.c',
+        'src/turbohtml/_c/css/select/to_xpath.c',
         'src/turbohtml/_c/query/xslt.c',
         'src/turbohtml/_c/query/methods.c',
         'src/turbohtml/_c/query/xpath/ast.c',
@@ -97,7 +97,7 @@ py.extension_module(
         'src/turbohtml/_c/extract/feed.c',
         'src/turbohtml/_c/extract/tables.c',
         'src/turbohtml/_c/extract/dates.c',
-        'src/turbohtml/_c/cssom/cssom.c',
+        'src/turbohtml/_c/css/cssom/cssom.c',
         'src/turbohtml/_c/validate/schema.c',
         'src/turbohtml/_c/validate/conformance.c',
     ],
