+++ src/turbohtml/_c/query/methods.c
@@ -6,7 +6,7 @@
 
 #include "core/vec.h"
 #include "dom/nodes.h"
-#include "query/css/selector.h"
+#include "css/select/selector.h"
 
 void handle_clear_caches(HandleObject *handle) {
     for (int index = 0; index < handle->sel_cache_len; index++) {
